+++ tests/test_mousemanager.py
@@ -19,8 +19,6 @@
 # 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  US
 
 
-from __future__ import print_function
-
 import os
 import sys