+++ tests/test_widgets.py
@@ -18,8 +18,6 @@
 # with this module if not, write to the Free Software Foundation, Inc.,
 # 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  US
 
-from __future__ import print_function
-
 import os
 import sys
 import logging